+++ src/query/sql/src/planner/optimizer/rule/rewrite/mod.rs
@@ -18,7 +18,6 @@ mod rule_commute_join;
 mod rule_eliminate_eval_scalar;
 mod rule_eliminate_filter;
 mod rule_eliminate_sort;
-mod rule_eliminate_subquery;
 mod rule_eliminate_union;
 mod rule_filter_nulls;
 mod rule_fold_count_aggregate;
@@ -54,7 +53,6 @@ pub use rule_commute_join::RuleCommuteJoin;
 pub use rule_eliminate_eval_scalar::RuleEliminateEvalScalar;
 pub use rule_eliminate_filter::RuleEliminateFilter;
 pub use rule_eliminate_sort::RuleEliminateSort;
-pub use rule_eliminate_subquery::RuleEliminateSubquery;
 pub use rule_eliminate_union::RuleEliminateUnion;
 pub use rule_filter_nulls::RuleFilterNulls;
 pub use rule_fold_count_aggregate::RuleFoldCountAggregate;
