+++ src/query/sql/src/planner/optimizer/rule/rule.rs
@@ -58,8 +58,6 @@ pub static DEFAULT_REWRITE_RULES: LazyLock<Vec<RuleID>> = LazyLock::new(|| {
         RuleID::PushDownFilterScan,
         RuleID::PushDownPrewhere, /* PushDownPrwhere should be after all rules except PushDownFilterScan */
         RuleID::PushDownSortScan, // PushDownSortScan should be after PushDownPrewhere
-        // Subquery
-        RuleID::EliminateSubquery,
     ]
 });
 
@@ -108,7 +106,6 @@ pub enum RuleID {
     EliminateEvalScalar,
     EliminateFilter,
     EliminateSort,
-    EliminateSubquery,
     MergeEvalScalar,
     MergeFilter,
     SplitAggregate,
@@ -168,7 +165,6 @@ impl Display for RuleID {
             RuleID::EliminateUnion => write!(f, "EliminateUnion"),
 
             RuleID::MergeFilterIntoMutation => write!(f, "MergeFilterIntoMutation"),
-            RuleID::EliminateSubquery => write!(f, "SubqueryEliminateLeftScan"),
         }
     }
 }
